+++ packages/mel-components/src/mel-select/demo/使用 optionsApi 传递请求数据源的参数.vue
@@ -40,8 +40,7 @@
 
   function getGoods() {
     return axios
-      .get<ResponseData<SelectorItem[]>>('https://apifoxmock.com/m1/4781098-4434938-default/selector/foods')
-      .then(res => res.data.data);
+      .get<ResponseData<SelectorItem[]>>('https://apifoxmock.com/m1/4781098-4434938-default/selector/foods');
   }
 
   watchImmediate(immediate, (immediate) => {